+++ src/main/java/com/realworld/project/application/user/service/CredentialService.java
@@ -1,6 +1,7 @@
 package com.realworld.project.application.user.service;
 
 import com.realworld.project.application.user.api.dto.RequestLoginUser;
+import com.realworld.project.application.user.api.dto.RequestRegisterUser;
 import com.realworld.project.application.user.api.dto.ResponseUser;
 import com.realworld.project.application.user.domain.Users;
 import com.realworld.project.application.user.repository.UsersRepository;
@@ -8,12 +9,13 @@ import com.realworld.project.core.jwt.JwtTokenProvider;
 import com.realworld.project.core.exception.BusinessException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import static com.realworld.project.core.exception.ErrorCode.USER_NOT_FOUND;
 
 @RequiredArgsConstructor
 @Service
-public class AuthService {
+public class CredentialService {
 
     private final UsersRepository usersRepository;
     private final JwtTokenProvider jwtTokenProvider;
@@ -25,5 +27,12 @@ public class AuthService {
         String token = jwtTokenProvider.createToken(Long.toString(user.getId()), null);
 
         return ResponseUser.of(user, token);
+    }
+
+    @Transactional
+    public ResponseUser registerUsers(RequestRegisterUser requestRegisterUser) {
+        Users user = usersRepository.save(requestRegisterUser.toEntity());
+
+        return ResponseUser.from(user);
     }
 }
