+++ src/test/java/com/realworld/project/application/user/service/CredentialServiceTest.java
@@ -1,6 +1,7 @@
 package com.realworld.project.application.user.service;
 
 import com.realworld.project.application.user.api.dto.RequestLoginUser;
+import com.realworld.project.application.user.api.dto.RequestRegisterUser;
 import com.realworld.project.application.user.api.dto.ResponseUser;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -12,19 +13,17 @@ import org.springframework.transaction.annotation.Transactional;
 import static com.realworld.project.fixture.UserFixture.*;
 import static org.assertj.core.api.Assertions.*;
 
+@DisplayName("회원자격관련 서비스")
 @Transactional
 @SpringBootTest
-class AuthServiceTest {
+class CredentialServiceTest {
 
     @Autowired
-    AuthService authService;
-
-    @Autowired
-    UserService userService;
+    CredentialService credentialService;
 
     @BeforeEach
     void before() {
-        register(userService);
+        register(credentialService);
     }
 
     @DisplayName("로그인")
@@ -33,11 +32,22 @@ class AuthServiceTest {
         //given
         RequestLoginUser requestLoginUser = new RequestLoginUser(email, password);
         //when
-        ResponseUser responseUser = authService.login(requestLoginUser);
+        ResponseUser responseUser = credentialService.login(requestLoginUser);
         //then
         assertThat(responseUser.getEmail()).isEqualTo(email);
         assertThat(responseUser.getUsername()).isEqualTo(username);
         assertThat(responseUser.getToken()).isNotNull();
     }
 
+    @DisplayName("회원가입")
+    @Test
+    void registerUsers() {
+        //given
+        RequestRegisterUser requestRegisterUser = new RequestRegisterUser(username, email, password);
+        //when
+        ResponseUser responseUser = credentialService.registerUsers(requestRegisterUser);
+        //then
+        assertThat(responseUser.getUsername()).isEqualTo(username);
+        assertThat(responseUser.getEmail()).isEqualTo(email);
+    }
 }
