+++ src/test/java/com/realworld/project/application/user/service/PasswordHelperTest.java
@@ -14,6 +14,7 @@ import static com.realworld.project.fixture.UserFixture.password;
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
 
+@DisplayName("패스워드 HelperClass ")
 @SpringBootTest
 class PasswordHelperTest {
